// src/app/modules/phan-he-chuyen-khoa/models/QuanLyDichVuModel.ts
export interface IXetNghiem {
  tenXetNghiem?: string;
  soLuong?: string;
  ketQua?: string;
  loaiBenhhPham?: string;
  doiTuong?: string;
}

export interface ICDHA {
  tenDichVu?: string;
  soLuong?: string;
  ghiChu?: string;
  ketLuan?: string;
  doiTuong?: string;
}

export interface IChuyenKhoa {
  tenDichVu?: string;
  sl?: string;
  loaiPTTT?: string;
  ghiChu?: string;
  batDau?: string;
  ketThuc?: string;
  ketLuan?: string;
  doiTuong?: string;
}

export interface IDVKhac {
  stt?: string;
  tenDichVu?: string;
  soLuong?: string;
  ghiChu?: string;
}